Rename files to `.js`
wp-scripts doesn't support `.jsx` extensions
