+++ public_html/wp-content/plugins/wcpt/javascript/tracker/source/components/filterable-table/header-cell.js
@@ -1,5 +1,5 @@
 import React, { PropTypes } from 'react';
-import SortingIndicator     from '../sorting-indicator/sorting-indicator.jsx';
+import SortingIndicator     from '../sorting-indicator/sorting-indicator';
 
 export default React.createClass( {
 	propTypes : {
