+++ public_html/wp-content/plugins/wcpt/javascript/tracker/source/components/filterable-table/header.js
@@ -1,5 +1,5 @@
 import React, { PropTypes } from 'react';
-import TableHeaderCell      from './header-cell.jsx';
+import TableHeaderCell      from './header-cell';
 
 export default React.createClass( {
 	propTypes : {
