+++ public_html/wp-content/plugins/wcpt/javascript/tracker/source/components/filterable-table/row.js
