+++ public_html/wp-content/plugins/wcpt/javascript/tracker/source/components/search-field/search-field.js
