+++ public_html/wp-content/plugins/wcpt/javascript/tracker/source/components/sorting-indicator/sorting-indicator.js
